+++ source/Public/Compare-ResourcePropertyState.ps1
@@ -131,6 +131,7 @@
 #>
 function Compare-ResourcePropertyState
 {
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseSyntacticallyCorrectExamples', '', Justification = 'Because the rule does not yet support parsing the code when a parameter type is not available. The ScriptAnalyzer rule UseSyntacticallyCorrectExamples will always error in the editor due to https://github.com/indented-automation/Indented.ScriptAnalyzerRules/issues/8.')]
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable[]])]
     param
@@ -202,7 +203,7 @@ function Compare-ResourcePropertyState
             }
     }
 
-    $compareTargetResourceStateReturnValue = @()
+    $compareTargetResourceStateReturnValue = [System.Collections.ArrayList]::new()
 
     foreach ($parameterName in $DesiredValues.Keys)
     {